- remove Perl section which disables warnings and uses blib
For some odd reason, make test would not start the server with this section. I
don't currently have any idea why, even though I was able to start the server
in single user mode, and under -debug using gdb and the Perl debugger.
Regardless, things are finally moving forward here, that was a head trip of a
bug.